Fixes #1859: Implemented support for line breaks within CSV fields
